requirements.txt
# Flask and Web Framework
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-Session==0.5.0
python-socketio==5.9.0

# LangChain Core
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.0.10

# LangChain LLM Integrations
langchain-openai==0.0.5
langchain-qianfan==0.0.1
langchain-deepseek==0.0.1

# LangChain Extras
langchain-experimental==0.0.50

# Data Models and Validation
pydantic==2.5.0

# Scientific Computing
numpy==1.24.3

# HTTP Requests
requests==2.31.0

# Configuration Management
PyYAML==6.0.1
python-dotenv==1.0.0

# Logging and Monitoring
psutil==5.9.6

# Development and Testing
pytest==7.4.3
pytest-flask==1.3.0

# Production (Optional)
gunicorn==21.2.0 